Improved implementation

+++ common/common.h
@@ -3,6 +3,6 @@
 
 #include <stdio.h>
 
-extern void log_data(FILE* stream, unsigned char* buf, unsigned int len);
+extern void log_data(FILE* stream, unsigned char* buf, size_t len);
 
 #endif
